pyproject.toml
[project]
name = "simple-dify-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "line-bot-sdk>=3.14.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "slack-bolt>=1.21.2",
]
